+++ Assets/Plugins/iOS/UAUnityMessageViewController.h
@@ -10,7 +10,7 @@
 #import "UAMessageCenterResources.h"
 #import "UAMessageCenterLocalization.h"
 #else
-@import Airship;
+@import AirshipKit;
 #endif
 
 NS_ASSUME_NONNULL_BEGIN