+++ Assets/Scripts/UrbanAirshipBehaviour.cs
@@ -10,8 +10,6 @@ public class UrbanAirshipBehaviour : MonoBehaviour {
     public string addTagOnStart;
 
     void Awake () {
-        UAirship.Shared.DataCollectionEnabled = true;
-        UAirship.Shared.PushTokenRegistrationEnabled = true;
         UAirship.Shared.UserNotificationsEnabled = true;
     }
 
@@ -19,6 +17,9 @@ public class UrbanAirshipBehaviour : MonoBehaviour {
         if (!string.IsNullOrEmpty (addTagOnStart)) {
             UAirship.Shared.AddTag (addTagOnStart);
         }
+
+        string[] allenable = new string[] { "FEATURE_ALL" };
+        UAirship.Shared.SetEnabledFeatures(allenable);
 
         UAirship.Shared.OnPushReceived += OnPushReceived;
         UAirship.Shared.OnChannelUpdated += OnChannelUpdated;